requirements.txt
streamlit>=1.36.0
transformers>=4.40.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python-headless>=4.8.0.74
pillow>=10.2.0
numpy>=1.24.0
pandas>=2.0.0